requirements.txt
Django==4.2.0
djangorestframework==3.14.0
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
openai==1.3.0
google-generativeai==0.3.0
Werkzeug==3.0.0
